fix: detect built-in tools when loading from assemble output
Using the --assemble flag will output the contents that can be read
again by gptscript to construct a tool without going through all the
steps again. This doesn't detect built-in tools, and will, instead, try
to run the built-in tools like they are commands.

This change will detect the built-in tools to ensure they are run
correctly.

Signed-off-by: Donnie Adams <donnie@acorn.io>