fix: the tool name was not used in the prompt
When the next tool chats, the name is lost when the resp variable is
reinitialized in the for loop. It always appeared to the user they
were chatting with the main entrypoint tool.

This changes it so the previous response is saved, and passed to the
set prompt function. So now if a tool is communicating the user
knows which tool is handling the chat turn.

Signed-off-by: Bill Maxwell <cloudnautique@users.noreply.github.com>