fix: do not lowercase all items in tool(like) lists (#344)
* fix: do not lowercase all items in tool(like) lists

When importing things from github or other case sensitive location
the previous functionality would lowercase everything causing
clones from VCS to fail.

This fix removes the calls to lowercase tools and contexts along with
the other places they occur.

Signed-off-by: Bill Maxwell <cloudnautique@users.noreply.github.com>

* fix: the tool name was not used in the prompt

When the next tool chats, the name is lost when the resp variable is
reinitialized in the for loop. It always appeared to the user they
were chatting with the main entrypoint tool.

This changes it so the previous response is saved, and passed to the
set prompt function. So now if a tool is communicating the user
knows which tool is handling the chat turn.

Signed-off-by: Bill Maxwell <cloudnautique@users.noreply.github.com>

---------

Signed-off-by: Bill Maxwell <cloudnautique@users.noreply.github.com>