fix: ensure fd monitor file is closed at end of the script run
When running gptscript with providers and --events-stream-to flag, the
file passed to --events-stream-to would be closed after the provider was
setup. This is an error because the file should remain open for the
duration of the script run.

This change will ensure that the file is only closed when the script is
complete.

Additionally, this change introduces 0644 permissions when creating a
file passed to --events-stream-to.

Signed-off-by: Donnie Adams <donnie@acorn.io>