fix: do not add readOnly fields to the requestBody
According to the spec readOnly fields SHOULD NOT be sent in
the request, but maybe present in the response. The LLM has made
several errors making calls when these fields are present by including
the readOnly fields in the request and getting HTTP 422's.

This change loops through the args, and drops readOnly fields from the
requestBodyContent arg on the tool.

Signed-off-by: Bill Maxwell <cloudnautique@users.noreply.github.com>