bug: go back to old behavior of dropping {} for tools
This is probably the wrong thing to do. We need to differientiate
between empty args and no args. If a tool has no params defined
then {} should be dropped. Right now this logic is in the wrong spot
to do that so we just go back to the old behavior of just always dropping
{} which will cause a problem for tools that would expect an empty object.