fix: ensure default model is parsed when using default model provider
When using the default model provider flag, the default model will be
gpt-4o. However, when it is parsed, the expected format is 'gpt-4p from
my-model-provider'. When using a default model provider, the 'gpt-4o'
was being parsed and the model was set to the empty string.

After this change, if the model name is empty after parsing, then we
know that the model from the request doesn't have a 'from' in it and the
first part of the split reference should be used.

Signed-off-by: Donnie Adams <donnie@acorn.io>