feat: allow providers to be restarted if they stop
By not caching the client, gptscript is able to restart the provider
daemon if it stops. If the daemon is still running, then there is little
overhead because the daemon URL is cached and the tool will not be
completely reprocessed.

The model to provider mapping is still cached so that the client can be
recreated when necessary.

Signed-off-by: Donnie Adams <donnie@acorn.io>