feat: add workspace API to SDK (#872)
This change also changes the error behavior when running tools that are
simply wrapped commands. Previously, all such tools would not return an
error, rather an error message in hopes that the LLM would retry.
However, if the tool is just a command (i.e. has no parent), then it
should return an error so that the caller doesn't have to guess whether
an error occurred.

Signed-off-by: Donnie Adams <donnie@acorn.io>