enhance: clear cache on load with disabled cache
A change was made such that the cache key was removed when a tool was
run with cache disabled. This change connects the cache disabled feature
when running tools with the cache disabled when loading tools. That way,
if a tool is loaded with cache disabled, then this newer version is used
when running the tool the next time.

Signed-off-by: Donnie Adams <donnie@acorn.io>