feat: add ability to gracefully stop runs (#950)
Various parts of a tool execution cannot be stopped gracefully. For example,
non-streamed HTTP request can't be stopped gracefully. However, commands and
chat completions can be gracefully stopped by a user and the result returned.
An "ABORTED BY USER" message is added to such messages. Additionally, aborted
chat completion responses are not stored in the cache.

Signed-off-by: Donnie Adams <donnie@acorn.io>