chore: switch from gob to JSON encoding for digests
The gob encoder is dependent on the order in which it sees types. So,
even encoding a completely new type could change the "digest" of a type
with the same values. This is not in the flavor or producing a digest.

This change switches to JSON, which will produce a consistent encoding.

Signed-off-by: Donnie Adams <donnie@acorn.io>